+++ src/df-architect.ts
@@ -1,10 +1,11 @@
 import DECLARATIONS from './core/_declarations.js'; DECLARATIONS();
 import ARCHITECT from "./core/architect.js";
+import HotkeyConfig from './forms/HotkeySettings.js';
 import { HOTKEYS } from "./core/hotkeys.js";
 import { LayerShortcuts } from "./general/LayerShortcuts.js";
 import { AltGridSnap } from './general/AltGridSnap.js';
 import { WallCtrlInvert } from './walls/WallCtrlInvert.js';
-import HotkeyConfig from './forms/HotkeySettings.js';
+import { WallShortcuts } from './walls/WallShortcuts.js';
 
 
 Hooks.once('init', function () {
@@ -13,6 +14,7 @@ Hooks.once('init', function () {
 	HotkeyConfig.init();
 	AltGridSnap.init();
 	WallCtrlInvert.init();
+	WallShortcuts.init();
 });
 
 Hooks.once('setup', function () { });