+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Layer Hot Swap shortcut. You can select 2 layers that a hotkey can quickly swap between. Defaulted to `Alt + Q`.
 - Added Wall Chain Lock. This essentially inverts the Ctrl key behaviour.
 - Added Wall Chain Lock hotkey, defaulted to `Alt + C`.
+- Added Wall Type shortcuts. These are defaulted to the keys `Shift + 1` to `Shift + 6`.
 
 ## Alpha Release v0.2.0 (2021-03-07)
 - Alternate Grid Snapping now works perfectly.